# Improve performance of looking up table entries by index