--- a/test/benchmark-common-tasks/main.cpp
+++ b/test/benchmark-common-tasks/main.cpp
@@ -1703,6 +1703,74 @@ struct BenchmarkInitiatorOpen : public BenchmarkNonInitiatorOpen {
     }
 };
 
+
+struct IterateTableByIndexNoPrimaryKey : Benchmark {
+    const char* name() const override
+    {
+        return "IterateTableByIndexNoPrimaryKey";
+    }
+
+    static const int row_count = 100'000;
+
+    void before_all(DBRef db) override
+    {
+        WrtTrans tr(db);
+        TableRef t = tr.add_table(name());
+        for (int i = 0; i < row_count; ++i)
+            t->create_object();
+        tr.commit();
+
+        m_tr.reset(new WrtTrans(db));
+        m_table = m_tr->get_table(name());
+    }
+    void before_each(DBRef) override {}
+    void after_each(DBRef) override {}
+
+    void operator()(DBRef) override
+    {
+        for (size_t i = 0, size = m_table->size(); i < size; ++i)
+            m_table->get_object(i);
+    }
+};
+
+struct IterateTableByIndexIntPrimaryKey : IterateTableByIndexNoPrimaryKey {
+    const char* name() const override
+    {
+        return "IterateTableByIndexIntPrimaryKey";
+    }
+
+    void before_all(DBRef db) override
+    {
+        WrtTrans tr(db);
+        TableRef t = tr.get_group().add_table_with_primary_key("class_table", type_Int, "pk", false);
+        for (int i = 0; i < row_count; ++i)
+            t->create_object_with_primary_key(i);
+        tr.commit();
+
+        m_tr.reset(new WrtTrans(db));
+        m_table = m_tr->get_table("class_table");
+    }
+};
+
+struct IterateTableByIndexStringPrimaryKey : IterateTableByIndexNoPrimaryKey {
+    const char* name() const override
+    {
+        return "IterateTableByIndexStringPrimaryKey";
+    }
+
+    void before_all(DBRef db) override
+    {
+        WrtTrans tr(db);
+        TableRef t = tr.get_group().add_table_with_primary_key("class_table", type_String, "pk", false);
+        for (int i = 0; i < row_count; ++i)
+            t->create_object_with_primary_key(std::to_string(i));
+        tr.commit();
+
+        m_tr.reset(new WrtTrans(db));
+        m_table = m_tr->get_table("class_table");
+    }
+};
+
 const char* to_lead_cstr(RealmDurability level)
 {
     switch (level) {
@@ -1847,6 +1915,10 @@ int benchmark_common_tasks_main()
     BENCH2(BenchmarkInitiatorOpen, true);
     BENCH2(AddTable, true);
     BENCH2(AddTable, false);
+
+    BENCH(IterateTableByIndexNoPrimaryKey);
+    BENCH(IterateTableByIndexIntPrimaryKey);
+    BENCH(IterateTableByIndexStringPrimaryKey);
 
     BENCH(BenchmarkSort);
     BENCH(BenchmarkSortInt);